geometry-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pythonocc-core==7.7.2
gunicorn==21.2.0
